changing parse

+++ public/js/kpiStarts.app.js
@@ -24,7 +24,7 @@ var kpiStartsApp = new Vue({
     formatSensorTime(){
       this.sensorTime.forEach(
         (entry, index, arr) => {
-          entry.starts = parseInt(entry.starts);
+          entry.starts = parseFloat(entry.starts);
           entry.dateCollected = Date.parse(entry.dataCollectedDate);
         }
       )
